+++ sentry-kotlin-multiplatform/src/appleMain/kotlin/io/sentry/kotlin/multiplatform/extensions/SentryOptionsExtensions.apple.kt
@@ -22,6 +22,9 @@ internal fun SentryOptions.toCocoaOptionsConfiguration(): (CocoaSentryOptions?) 
 internal fun CocoaSentryOptions.applyCocoaBaseOptions(kmpOptions: SentryOptions) {
     val cocoaOptions = this
     cocoaOptions.setDsn(kmpOptions.dsn)
+    cocoaOptions.setStrictTraceContinuation(kmpOptions.strictTraceContinuation)
+    cocoaOptions.setOrgId(kmpOptions.orgId)
+    cocoaOptions.setEnablePropagateTraceparent(kmpOptions.enablePropagateTraceparent)
     cocoaOptions.setAttachStacktrace(kmpOptions.attachStackTrace)
     cocoaOptions.setDist(kmpOptions.dist)
     kmpOptions.environment?.let {