+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/extensions/SentryOptionsExtensions.jvm.kt
@@ -22,6 +22,9 @@ internal fun SentryOptions.toJvmSentryOptionsCallback(): (JvmSentryOptions) -> U
 internal fun JvmSentryOptions.applyJvmBaseOptions(kmpOptions: SentryOptions) {
     val jvmOptions = this
     jvmOptions.dsn = kmpOptions.dsn
+    jvmOptions.isStrictTraceContinuation = kmpOptions.strictTraceContinuation
+    jvmOptions.orgId = kmpOptions.orgId
+    jvmOptions.isPropagateTraceparent = kmpOptions.enablePropagateTraceparent
     jvmOptions.isAttachThreads = kmpOptions.attachThreads
     jvmOptions.isAttachStacktrace = kmpOptions.attachStackTrace
     jvmOptions.dist = kmpOptions.dist